Fix type errors

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movement/service/mapper/MovementModelToEntityMapper.java
@@ -124,7 +124,7 @@ public class MovementModelToEntityMapper {
                 entity.setLocation(point);
                 
                 if (movement.getPosition().getAltitude() != null) {
-                    entity.setAltitude(movement.getPosition().getAltitude().intValue());
+                    entity.setAltitude(movement.getPosition().getAltitude());
                 }
             }
